refactored ui
