+++ PowerPointLabs/PowerPointLabs/ImageSearch/ImageSearchPane.xaml.cs
@@ -1,38 +1,32 @@
 ﻿using System;
 using System.Collections.Generic;
 using System.Collections.ObjectModel;
+using System.Collections.Specialized;
 using System.ComponentModel;
-using System.Drawing;
-using System.Linq;
 using System.Windows;
 using System.Windows.Controls;
 using System.Windows.Controls.Primitives;
-using System.Windows.Forms;
 using System.Windows.Input;
-using MahApps.Metro.Controls.Dialogs;
-using PowerPointLabs.AutoUpdate;
 using PowerPointLabs.ImageSearch.Domain;
 using PowerPointLabs.ImageSearch.Handler;
 using PowerPointLabs.ImageSearch.SearchEngine;
-using PowerPointLabs.ImageSearch.SearchEngine.VO;
 using PowerPointLabs.ImageSearch.Util;
 using PowerPointLabs.Models;
 using PowerPointLabs.WPF.Observable;
 using ButtonBase = System.Windows.Controls.Primitives.ButtonBase;
-using Clipboard = System.Windows.Clipboard;
-using Image = System.Drawing.Image;
 using KeyEventArgs = System.Windows.Input.KeyEventArgs;
 using ListBox = System.Windows.Controls.ListBox;
 using Timer = System.Timers.Timer;
 
 namespace PowerPointLabs.ImageSearch
 {
-    // TODO more partial class to clean up codes
     /// <summary>
     /// Interaction logic for ImageSearchPane.xaml
     /// </summary>
     public partial class ImageSearchPane
     {
+        # region Props & States
+
         // UI model - list that holds search result item
         public ObservableCollection<ImageItem> SearchList { get; set; }
 
@@ -77,17 +71,10 @@ namespace PowerPointLabs.ImageSearch
         private readonly HashSet<string> _insertDownloadingUriList = new HashSet<string>();
         private readonly Dictionary<string, ImageItem> _insertDownloadingUriToPreviewImage = new Dictionary<string, ImageItem>();
 
-        // TODO put to text collection
-        private const string ErrorNetworkOrApiQuotaUnavailable =
-            "Failed to search images. Please check your network, or the daily API quota is ran out.";
-
-        private const string ErrorNetworkOrSourceUnavailable =
-            "Failed to insert style. Please check your network, or the image source is unavailable.";
-
-        private const string ErrorNoEngineIdOrApiKey =
-            "Please fill in Search Engine Id and API Key by clicking ADVANCED.. button.";
+        # endregion
 
         #region Initialization
+
         public ImageSearchPane()
         {
             InitializeComponent();
@@ -95,14 +82,10 @@ namespace PowerPointLabs.ImageSearch
             InitMultiplePurposeButtons();
             InitSearchList();
             InitPreviewList();
-            SearchListBox.DataContext = this;
-            PreviewListBox.DataContext = this;
             IsOpen = true;
             InitStyleOptions();
             InitSearchOptions();
-
-            var isTempFolderReady = TempPath.InitTempFolder();
-            if (isTempFolderReady)
+            if (TempPath.InitTempFolder())
             {
                 InitSearchEngine();
                 InitPreviewPresentation();
@@ -112,13 +95,18 @@ namespace PowerPointLabs.ImageSearch
 
         private void InitSearchTextbox()
         {
-            SearchTextboxWatermark = new ObservableString {Text = "Type here to search for images"};
+            SearchTextboxWatermark = new ObservableString {Text = TextCollection.ImagesLabText.TextBoxWatermarkSearch};
             SearchTextBox.DataContext = SearchTextboxWatermark;
         }
 
         private void InitMultiplePurposeButtons()
         {
-            MultiplePurposeButtons = new ObservableCollection<string>(new List<string> { "Search", "Download", "From File" });
+            MultiplePurposeButtons = new ObservableCollection<string>(new List<string>
+            {
+                TextCollection.ImagesLabText.MultiPurposeButtonNameSearch, 
+                TextCollection.ImagesLabText.MultiPurposeButtonNameDownload, 
+                TextCollection.ImagesLabText.MultiPurposeButtonNameFromFile
+            });
             SearchButton.ItemsSource = MultiplePurposeButtons;
             SearchButton.SelectedIndex = 0;
         }
@@ -126,24 +114,8 @@ namespace PowerPointLabs.ImageSearch
         private void InitPreviewList()
         {
             PreviewList = new ObservableCollection<ImageItem>();
-            PreviewList.CollectionChanged += (sender, args) =>
-            {
-                if (PreviewList.Count != 0)
-                {
-                    PreviewInstructions.Visibility = Visibility.Hidden;
-                    PreviewInstructionsWhenNoSelectedSlide.Visibility = Visibility.Hidden;
-                }
-                else if (PowerPointCurrentPresentationInfo.CurrentSlide == null)
-                {
-                    PreviewInstructionsWhenNoSelectedSlide.Visibility = Visibility.Visible;
-                    PreviewInstructions.Visibility = Visibility.Hidden;
-                }
-                else
-                {
-                    PreviewInstructions.Visibility = Visibility.Visible;
-                    PreviewInstructionsWhenNoSelectedSlide.Visibility = Visibility.Hidden;
-                }
-            };
+            PreviewList.CollectionChanged += PreviewList_OnCollectionChanged;
+            PreviewListBox.DataContext = this;
         }
 
         private void InitSearchList()
@@ -151,12 +123,8 @@ namespace PowerPointLabs.ImageSearch
             SearchList = new ObservableCollection<ImageItem>();
             _downloadedImages = new List<ImageItem>();
             _fromFileImages = new List<ImageItem>();
-            SearchList.CollectionChanged += (sender, args) =>
-            {
-                SearchInstructions.Visibility = SearchList.Count == 0 
-                    ? Visibility.Visible 
-                    : Visibility.Hidden;
-            };
+            SearchList.CollectionChanged += SearchList_OnCollectionChanged;
+            SearchListBox.DataContext = this;
         }
 
         private void InitSearchOptions()
@@ -169,141 +137,7 @@ namespace PowerPointLabs.ImageSearch
         {
             StyleOptions = StyleOptions.Load(StoragePath.GetPath("ImagesLabStyleOptions"));
             OptionsPane.DataContext = StyleOptions;
-            StyleOptionsFlyout.IsOpenChanged += (sender, args) =>
-            {
-                if (!StyleOptionsFlyout.IsOpen)
-                {
-                    DoPreview();
-                }
-            };
-        }
-
-        private void InitSearchEngine()
-        {
-            SearchEngine = new GoogleEngine(SearchOptions)
-                .WhenSucceed(WhenSearchSucceed())
-                .WhenCompleted(WhenSearchCompleted())
-                .WhenFail(response => {
-                    ShowErrorMessageBox(ErrorNetworkOrApiQuotaUnavailable);
-                    Dispatcher.BeginInvoke(new Action(() =>
-                    {
-                        // if all failed
-                        if (SearchList.Count(source => source.ImageFile == TempPath.LoadingImgPath) 
-                            == SearchList.Count)
-                        {
-                            SearchList.Clear();
-                        }
-                    }));
-                });
-        }
-
-        private void ShowErrorMessageBox(string content)
-        {
-            Dispatcher.BeginInvoke(new Action(() =>
-            {
-                this.ShowMessageAsync("Error", content);
-            }));
-        }
-
-        private GoogleEngine.WhenCompletedEventDelegate WhenSearchCompleted()
-        {
-            return isSuccess =>
-            {
-                Dispatcher.BeginInvoke(new Action(() =>
-                {
-                    SearchProgressRing.IsActive = false;
-                    var isThereMoreSearchResults = !RemoveElementInSearchList(item => item.IsToDelete);
-                    SearchButton.IsEnabled = true;
-                    if (isSuccess 
-                        && isThereMoreSearchResults 
-                        && SearchList.Count + GoogleEngine.NumOfItemsPerRequest - 1/*loadMore item*/ 
-                        <= GoogleEngine.MaxNumOfItems)
-                    {
-                        SearchList.Add(new ImageItem
-                        {
-                            ImageFile = TempPath.LoadMoreImgPath
-                        });
-                    }
-                }));
-            };
-        }
-
-        // TODO util
-        private bool RemoveElementInSearchList(Func<ImageItem, bool> cond)
-        {
-            var isAnyElementRemoved = false;
-            for (var i = 0; i < SearchList.Count;)
-            {
-                if (cond(SearchList[i]))
-                {
-                    SearchList.RemoveAt(i);
-                    isAnyElementRemoved = true;
-                }
-                else
-                {
-                    i++;
-                }
-            }
-            return isAnyElementRemoved;
-        }
-
-        private GoogleEngine.WhenSucceedEventDelegate WhenSearchSucceed()
-        {
-            return (searchResults, startIdx) =>
-            {
-                searchResults.Items = searchResults.Items ?? new List<SearchResult>();
-                for (var i = 0; i < GoogleEngine.NumOfItemsPerRequest; i++)
-                {
-                    var item = SearchList[startIdx + i];
-                    if (i >= searchResults.Items.Count)
-                    {
-                        item.IsToDelete = true;
-                        continue;
-                    }
-
-                    var searchResult = searchResults.Items[i];
-                    var thumbnailPath = TempPath.GetPath("thumbnail");
-
-                    new Downloader()
-                        .Get(searchResult.Image.ThumbnailLink, thumbnailPath)
-                        .After(AfterDownloadThumbnail(item, thumbnailPath, searchResult))
-                        .Start();
-                }
-            };
-        }
-
-        private Downloader.AfterDownloadEventDelegate AfterDownloadThumbnail(
-            ImageItem item, string thumbnailPath, SearchResult searchResult)
-        {
-            return () =>
-            {
-                if (item == null) return;
-                item.ImageFile = thumbnailPath;
-                if (searchResult != null)
-                {
-                    item.FullSizeImageUri = searchResult.Link;
-                    item.Tooltip = GetTooltip(searchResult);
-                    item.ContextLink = searchResult.Image.ContextLink;
-                }
-                else // use case when thumbnail is already full-size
-                {
-                    item.FullSizeImageFile = item.ImageFile;
-                }
-
-                Dispatcher.BeginInvoke(new Action(() =>
-                {
-                    var selectedImageItem = SearchListBox.SelectedValue as ImageItem;
-                    if (selectedImageItem != null && item.ImageFile == selectedImageItem.ImageFile)
-                    {
-                        DoPreview();
-                    }
-                }));
-            };
-        }
-
-        private static string GetTooltip(SearchResult searchResult)
-        {
-            return searchResult.Title + "\n" + searchResult.Image.Width + " x " + searchResult.Image.Height;
+            StyleOptionsFlyout.IsOpenChanged += StyleOptionsFlyout_OnIsOpenChanged;
         }
 
         private void InitPreviewPresentation()
@@ -312,244 +146,56 @@ namespace PowerPointLabs.ImageSearch
             PreviewPresentation.Open(withWindow: false, focus: false);
         }
 
-        // intent:
-        // when select a thumbnail for some time (defined by TimerInterval),
-        // try to download its full size version for better preview and can be used for insertion
-        private void InitPreviewTimer()
-        {
-            PreviewTimer = new Timer { Interval = TimerInterval };
-            PreviewTimer.Elapsed += (sender, args) =>
-            {
-                // in timer thread
-                PreviewTimer.Stop();
-                Dispatcher.BeginInvoke(new Action(() =>
-                {
-                    // UI thread starts
-                    var source = SearchListBox.SelectedValue as ImageItem;
-                    // if already have cached full-size image, ignore
-                    if (source == null || source.FullSizeImageFile != null)
-                    {
-                        // do nothing
-                    }
-                    // if not downloading the full size image yet, download it
-                    else if (!_timerDownloadingUriList.Contains(source.FullSizeImageUri))
-                    {
-                        _timerDownloadingUriList.Add(source.FullSizeImageUri);
-                        // preview progress ring will be off, after preview processing is done
-                        PreviewProgressRing.IsActive = true;
-
-                        var fullsizeImageFile = TempPath.GetPath("fullsize");
-                        new Downloader()
-                            .Get(source.FullSizeImageUri, fullsizeImageFile)
-                            .After(AfterDownloadFullSizeImage(source, fullsizeImageFile))
-                            .OnError(WhenFailDownloadFullSizeImage())
-                            .Start();
-                    }
-                    // it's downloading
-                    else
-                    {
-                        // preview progress ring will be off, after preview processing is done
-                        PreviewProgressRing.IsActive = true;
-                    }
-                }));
-            };
-        }
-
-        private Downloader.ErrorEventDelegate WhenFailDownloadFullSizeImage()
-        {
-            return () =>
-            {
-                // in downloader thread
-                Dispatcher.BeginInvoke(new Action(() =>
-                {
-                    PreviewProgressRing.IsActive = false;
-                }));
-            };
-        }
-
-        private Downloader.AfterDownloadEventDelegate
-            AfterDownloadFullSizeImage(ImageItem source, string fullsizeImageFile)
-        {
-            // timer's downloading will come here at the end,
-            // or both timer + insert's downloading will come here
-            return () =>
-            {
-                // in downloader thread
-                Dispatcher.BeginInvoke(new Action(() =>
-                {
-                    if (source == null) return;
-
-                    // UI thread again
-                    // store back to image, so cache it
-                    source.FullSizeImageFile = fullsizeImageFile;
-                    var fullsizeImageUri = source.FullSizeImageUri;
-
-                    // intent: during download, selected item may have been changed to another one
-                    // if selected one got changed,
-                    // 1. no need to preview it
-                    // 2. no need to insert it to current slide
-                    var currentImageItem = SearchListBox.SelectedValue as ImageItem;
-                    if (currentImageItem == null)
-                    {
-                        PreviewProgressRing.IsActive = false;
-                    }
-                    else if (currentImageItem.ImageFile == source.ImageFile)
-                    {
-                        // if selected one remains
-                        // and it is to insert the full size image,
-                        ImageItem targetStyle;
-                        if (_insertDownloadingUriList.Contains(fullsizeImageUri)
-                            && _insertDownloadingUriToPreviewImage
-                                .TryGetValue(fullsizeImageUri, out targetStyle))
-                        {
-                            // insert + do preview
-                            PreviewPresentation.ApplyStyle(source, targetStyle.Tooltip);
-                            DoPreview(source);
-                        } 
-                        // or it is to preview only (from timer)
-                        else if (_timerDownloadingUriList.Contains(fullsizeImageUri))
-                        {
-                            DoPreview(source);
-                        }
-                    }
-
-                    RemoveDebounceCheck(fullsizeImageUri);
-                }));
-            };
-        }
-
-        private void RemoveDebounceCheck(string fullsizeImageUri)
-        {
-            if (_insertDownloadingUriList.Remove(fullsizeImageUri))
-            {
-                _insertDownloadingUriToPreviewImage.Remove(fullsizeImageUri);
-            }
-            _timerDownloadingUriList.Remove(fullsizeImageUri);
-        }
-
         # endregion
+
+        private void StyleOptionsFlyout_OnIsOpenChanged(object sender, RoutedEventArgs e)
+        {
+            if (!StyleOptionsFlyout.IsOpen)
+            {
+                DoPreview();
+            }
+        }
+
+        private void SearchList_OnCollectionChanged(object sender, NotifyCollectionChangedEventArgs e)
+        {
+            SearchInstructions.Visibility = SearchList.Count == 0
+                ? Visibility.Visible
+                : Visibility.Hidden;
+        }
+
+        private void PreviewList_OnCollectionChanged(object sender, NotifyCollectionChangedEventArgs e)
+        {
+            if (PreviewList.Count != 0)
+            {
+                PreviewInstructions.Visibility = Visibility.Hidden;
+                PreviewInstructionsWhenNoSelectedSlide.Visibility = Visibility.Hidden;
+            }
+            else if (PowerPointCurrentPresentationInfo.CurrentSlide == null)
+            {
+                PreviewInstructionsWhenNoSelectedSlide.Visibility = Visibility.Visible;
+                PreviewInstructions.Visibility = Visibility.Hidden;
+            }
+            else
+            {
+                PreviewInstructions.Visibility = Visibility.Visible;
+                PreviewInstructionsWhenNoSelectedSlide.Visibility = Visibility.Hidden;
+            }
+        }
 
         private void SearchButton_OnClick(object sender, RoutedEventArgs e)
         {
             switch (SearchButton.SelectedIndex)
             {
                 case 0:// search
-                    StartSearching();
+                    DoSearch();
                     break;
                 case 1:// download link
-                    var downloadLink = SearchTextBox.Text.Trim();
-                    if (downloadLink.Length == 0)
-                    {
-                        return;
-                    }
-
-                    // TODO parse downloadLink.. 
-                    // maybe a link from google image?
-                    // maybe without http://
-                    var item = new ImageItem
-                    {
-                        ImageFile = TempPath.LoadingImgPath,
-                        ContextLink = downloadLink
-                    };
-                    SearchList.Add(item);
-                    SearchProgressRing.IsActive = true;
-
-                    var thumbnailPath = TempPath.GetPath("thumbnail");
-                    new Downloader()
-                        .Get(downloadLink, thumbnailPath)
-                        .After(() =>
-                        {
-                            AfterDownloadThumbnail(item, thumbnailPath, null)();
-                            try
-                            {
-                                Image imgInput = Image.FromFile(thumbnailPath);
-                                Graphics.FromImage(imgInput);
-                                // so this is an image
-                                Dispatcher.Invoke(new Action(() =>
-                                {
-                                    SearchProgressRing.IsActive = false;
-                                    _downloadedImages.Add(item);
-                                }));
-                            }
-                            catch
-                            {
-                                // not an image
-                                Dispatcher.BeginInvoke(new Action(() =>
-                                {
-                                    SearchProgressRing.IsActive = false;
-                                    SearchList.Remove(item);
-                                }));
-                            }
-                        })
-                        .OnError(() =>
-                        {
-                            Dispatcher.BeginInvoke(new Action(() =>
-                            {
-                                SearchProgressRing.IsActive = false;
-                                SearchList.Remove(item);
-                            }));
-                            
-                        })
-                        .Start();
+                    DoDownloadImage();
                     break;
                 case 2:// image from file
-                    var openFileDialog = new OpenFileDialog
-                    {
-                        DefaultExt = "*.png",
-                        Multiselect = false,
-                        Filter = @"Image File|*.png;*.jpg;*.jpeg;"
-                    };
-                    var fileDialogResult = openFileDialog.ShowDialog();
-                    if (fileDialogResult != System.Windows.Forms.DialogResult.OK)
-                    {
-                        return;
-                    }
-
-                    var fromFileItem = new ImageItem
-                    {
-                        ImageFile = openFileDialog.FileName,
-                        FullSizeImageFile = openFileDialog.FileName,
-                        FullSizeImageUri = openFileDialog.FileName,
-                        ContextLink = openFileDialog.FileName
-                    };
-                    SearchList.Add(fromFileItem);
-                    _fromFileImages.Add(fromFileItem);
+                    DoLoadImageFromFile();
                     break;
             }
-        }
-
-        private void StartSearching()
-        {
-            var query = SearchTextBox.Text;
-            if (query.Trim().Length == 0)
-            {
-                return;
-            }
-            else if (SearchOptions.SearchEngineId.Trim().Length == 0
-                     || SearchOptions.ApiKey.Trim().Length == 0)
-            {
-                ShowErrorMessageBox(ErrorNoEngineIdOrApiKey);
-                return;
-            }
-
-            SearchButton.IsEnabled = false;
-            PrepareToSearch(GoogleEngine.NumOfItemsPerSearch);
-            SearchEngine.Search(query);
-        }
-
-        private void PrepareToSearch(int expectedNumOfImages, bool isListClearNeeded = true)
-        {
-            // clear search list, and show a list of
-            // 'Loading...' images
-            if (isListClearNeeded)
-            {
-                SearchList.Clear();
-            }
-            for (var i = 0; i < expectedNumOfImages; i++)
-            {
-                SearchList.Add(new ImageItem { ImageFile = TempPath.LoadingImgPath });
-            }
-            SearchProgressRing.IsActive = true;
         }
 
         // intent:
@@ -568,78 +214,24 @@ namespace PowerPointLabs.ImageSearch
         private void SearchListBox_OnSelectionChanged(object sender, SelectionChangedEventArgs e)
         {
             var source = (ImageItem) SearchListBox.SelectedValue;
-            if (source == null || source.ImageFile == TempPath.LoadingImgPath)
+            if (source != null && source.ImageFile == TempPath.LoadMoreImgPath)
             {
-                PreviewList.Clear();
-                PreviewProgressRing.IsActive = false;
-            } 
-            else if (source.ImageFile == TempPath.LoadMoreImgPath)
-            {
-                PreviewList.Clear();
-                PreviewProgressRing.IsActive = false;
-
-                SearchButton.IsEnabled = false;
-                source.ImageFile = TempPath.LoadingImgPath;
-                PrepareToSearch(GoogleEngine.NumOfItemsPerRequest - 1, isListClearNeeded: false);
-                SearchEngine.SearchMore();
+                DoSearchMore(source);
             }
             else
             {
-                // when selection changed, no need to insert style
-                if (_insertDownloadingUriList.Remove(source.FullSizeImageUri))
-                {
-                    _insertDownloadingUriToPreviewImage.Remove(source.FullSizeImageUri);
-                }
+                // when selection changed, no need to insert 
                 // but dont clear timerDownloadingUriList, since timer may still downloading
                 // full size image at the background.
-
-                PreviewTimer.Stop();
-
-                DoPreview(source);
-
-                // when timer ticks, try to download full size image to replace
-                PreviewTimer.Start();
+                DoPreview(() =>
+                {
+                    if (source != null && _insertDownloadingUriList.Remove(source.FullSizeImageUri))
+                    {
+                        _insertDownloadingUriToPreviewImage.Remove(source.FullSizeImageUri);
+                    }
+                });
             }
         }
-
-        // do preview processing
-        private void DoPreview(ImageItem source)
-        {
-            // ui thread
-            Dispatcher.BeginInvoke(new Action(() =>
-            {
-                var previousTextCopy = Clipboard.GetText();
-                var selectedId = PreviewListBox.SelectedIndex;
-                PreviewList.Clear();
-
-                if (PowerPointCurrentPresentationInfo.CurrentSlide != null)
-                {
-                    var previewInfo = PreviewPresentation.PreviewStyles(source);
-                    Add(PreviewList, previewInfo.DirectTextStyleImagePath, TextCollection.ImagesLabText.StyleNameDirectText);
-                    Add(PreviewList, previewInfo.BlurStyleImagePath, TextCollection.ImagesLabText.StyleNameBlur);
-                    Add(PreviewList, previewInfo.TextboxStyleImagePath, TextCollection.ImagesLabText.StyleNameTextBox);
-                    Add(PreviewList, previewInfo.GrayScaleStyleImagePath, TextCollection.ImagesLabText.StyleNameGrayscale);
-
-                    PreviewListBox.SelectedIndex = selectedId;
-                }
-                if (previousTextCopy.Length > 0)
-                {
-                    Clipboard.SetText(previousTextCopy);
-                }
-                PreviewProgressRing.IsActive = false;
-            }));
-        }
-
-        // TODO util
-        private void Add(ICollection<ImageItem> list, string imagePath, string tooltip)
-        {
-            list.Add(new ImageItem
-            {
-                ImageFile = imagePath,
-                Tooltip = tooltip
-            });
-        }
-
 
         // intent:
         // allow arrow keys to navigate the search result items in the list
@@ -674,7 +266,8 @@ namespace PowerPointLabs.ImageSearch
             }
 
             e.Handled = true;
-            var item = (ListBoxItem) listbox.ItemContainerGenerator.ContainerFromItem(listbox.SelectedItem);
+            var item = (ListBoxItem) listbox.ItemContainerGenerator
+                .ContainerFromItem(listbox.SelectedItem);
             item.Focus();
         }
 
@@ -689,16 +282,14 @@ namespace PowerPointLabs.ImageSearch
         // intent: drag splitter to change grid width
         private void Splitter_OnDragDelta(object sender, DragDeltaEventArgs e)
         {
-            ImagesLabGrid.ColumnDefinitions[0].Width = new GridLength(ImagesLabGrid.ColumnDefinitions[0].ActualWidth + e.HorizontalChange);
+            ImagesLabGrid.ColumnDefinitions[0].Width = 
+                new GridLength(ImagesLabGrid.ColumnDefinitions[0].ActualWidth + e.HorizontalChange);
         }
 
         // enable & disable insert button
         private void PreivewListBox_OnSelectionChanged(object sender, SelectionChangedEventArgs e)
         {
-            Dispatcher.BeginInvoke(new Action(() =>
-            {
-                PreviewInsert.IsEnabled = PreviewListBox.SelectedValue != null;
-            }));
+            PreviewInsert.IsEnabled = PreviewListBox.SelectedValue != null;
         }
 
         // rmb to close background presentation
@@ -713,39 +304,9 @@ namespace PowerPointLabs.ImageSearch
             SearchOptions.Save(StoragePath.GetPath("ImagesLabSearchOptions"));
         }
 
-        private void PreviewInsert_OnClick(object sender, RoutedEventArgs e)
+        private void PreviewApply_OnClick(object sender, RoutedEventArgs e)
         {
-            PreviewTimer.Stop();
-            PreviewProgressRing.IsActive = true;
-
-            var source = (ImageItem) SearchListBox.SelectedValue;
-            var targetStyle = (ImageItem) PreviewListBox.SelectedValue;
-            if (source == null || targetStyle == null) return;
-
-            if (source.FullSizeImageFile != null)
-            {
-                PreviewPresentation.ApplyStyle(source, targetStyle.Tooltip);
-                PreviewProgressRing.IsActive = false;
-            }
-            else if (!_insertDownloadingUriList.Contains(source.FullSizeImageUri))
-            {
-                var fullsizeImageUri = source.FullSizeImageUri;
-                _insertDownloadingUriList.Add(fullsizeImageUri);
-                _insertDownloadingUriToPreviewImage[fullsizeImageUri] = targetStyle;
-
-                var fullsizeImageFile = TempPath.GetPath("fullsize");
-                new Downloader()
-                    .Get(fullsizeImageUri, fullsizeImageFile)
-                    .After(AfterDownloadFullSizeImage(source, fullsizeImageFile))
-                    .OnError(() => { ShowErrorMessageBox(ErrorNetworkOrSourceUnavailable); })
-                    .Start();
-            }
-            // already downloading, then update preview image in the map
-            else
-            {
-                var fullsizeImageUri = source.FullSizeImageUri;
-                _insertDownloadingUriToPreviewImage[fullsizeImageUri] = targetStyle;
-            }
+            ApplyStyle();
         }
 
         private void PreviewDisplayToggleSwitch_OnIsCheckedChanged(object sender, EventArgs e)
@@ -757,22 +318,6 @@ namespace PowerPointLabs.ImageSearch
                     ? new GridLength(620) 
                     : new GridLength(320);
             }));
-        }
-
-        private void DoPreview()
-        {
-            var image = (ImageItem) SearchListBox.SelectedValue;
-            if (image == null || image.ImageFile == TempPath.LoadingImgPath)
-            {
-                PreviewList.Clear();
-                PreviewProgressRing.IsActive = false;
-            }
-            else
-            {
-                PreviewTimer.Stop();
-                DoPreview(image);
-                PreviewTimer.Start();
-            }
         }
 
         private void StyleOptionsButton_OnClick(object sender, RoutedEventArgs e)
@@ -810,28 +355,26 @@ namespace PowerPointLabs.ImageSearch
             SearchTextBox.Text = "";
             switch (SearchButton.SelectedIndex)
             {
-                // TODO remove those magic
                 case 0: // search
                     SearchTextBox.IsEnabled = true;
-                    SearchTextboxWatermark.Text = "Type here to search for images";
-                    SearchInstructions.Text = "No result. Type the keywords in the textbox above to search for images.";
+                    SearchTextboxWatermark.Text = TextCollection.ImagesLabText.TextBoxWatermarkSearch;
+                    SearchInstructions.Text = TextCollection.ImagesLabText.InstructionForSearch;
                     break;
                 case 1: // download
                     SearchTextBox.IsEnabled = true;
-                    SearchTextboxWatermark.Text = "Paste the image link here";
-                    SearchInstructions.Text = "No image. Paste the image link in the textbox above to download images.";
+                    SearchTextboxWatermark.Text = TextCollection.ImagesLabText.TextBoxWatermarkDownload;
+                    SearchInstructions.Text = TextCollection.ImagesLabText.InstructionForDownload;
                     CopyContentToObservableList(_downloadedImages, SearchList);
                     break;
                 case 2: // from file
                     SearchTextBox.IsEnabled = false;
-                    SearchTextboxWatermark.Text = "";
-                    SearchInstructions.Text = "No image. Click the 'From File' button above to load local images.";
+                    SearchTextboxWatermark.Text = TextCollection.ImagesLabText.TextBoxWatermarkFromFile;
+                    SearchInstructions.Text = TextCollection.ImagesLabText.InstructionForFromFile;
                     CopyContentToObservableList(_fromFileImages, SearchList);
                     break;
             }
         }
 
-        // TODO util
         private void CopyContentToObservableList(IEnumerable<ImageItem> content, ObservableCollection<ImageItem> target)
         {
             foreach (var image in content)
